Rephrase mention of Windows
The program might work on Windows Subsystem for Linux, or it might not –
I don’t know, and I have no way to find out. Rephrase the sentence to
more or less imply incompatibility with Windows (I am certain that this
is true for older Windows versions) while not directly stating something
possibly false or begging a reviewer question (“well does it work on
Windows or not?”).