Add screenshots of showcase item
Montblanc has been a showcase item [1] for years, and it has a few nice
properties: its first few statements demonstrate several different data
types; it does not have several “instance of” statements, which would
take up a lot of screen space; at the same time, it does have several
“official name” statements (a bit further down the list, and at the end
of the screenshot), showing that this is possible; the “official name”
statements also demonstrate qualifiers and references, while the earlier
statements (“instance of” and “image”) are simpler by lacking them.

All screenshots were taken on revision 704641959, which is the second
most recent revision of that page as of this writing, so that the edit
interface would not be loaded. A custom style sheet, found in
screenshots/wikidata-reduced.cess, was applied. The screenshot area is
195px/100px to 1320px/1020px, taken using the Firefox Screenshots tool.

This also includes a Spanish screenshot, even though it isn’t currently
used in the text, because I wasn’t sure if I should use Spanish or
Catalan. Perhaps I’ll change my mind later.

[1]: https://www.wikidata.org/wiki/Wikidata:Showcase_items