Use more glossary terms
In the grand scheme of things, this is a minor change. Its motivation is
that certain terms like “label”, “description” etc. should perhaps be
written in uppercase (Lydia requested this, I’m still skeptical), and
extracting them into glossary terms makes it easy to change that in a
single place. In the process, I also added a lot of other terms, since I
had to go through the whole text anyways.

Unfortunately, to Git this looks like a lot of changes, because a lot of
lines are touched, even though the change in each line is mostly minor.
If you found this commit through a `git blame`… I’m sorry. (Remember
that you can use `git blame PATH HASH^` to blame the file version of a
commit’s parent.)