Use old-style numerals for “RDF2Graph”
With the default font features in Linux Libertine, the “2” in
“RDF2Graph” is slightly less high than the capital letters surrounding
it, which I find a bit displeasing. Using old-style numerals instead
results in a markedly different “2” that is nowhere near the height of
the capital letters, setting it apart from them. This may also help to
emphasize the reading of “RDF-to-Graph” instead of “RDF2-Graph”.