Use \gls{RDF2Graph} in headings
When I added the glossary entries earlier, I wasn’t sure if using \gls{}
in headings would work or not, and couldn’t be bothered to try it out.
Turns out, it does work, no problem.

However, the glossary entry has to be slightly tweaked, so that only the
actual numeral is set in old-style numerals, not the rest of the word.
Otherwise, the entire word uses the default document font (serif, normal
weight) instead of the heading font (sans-serif, bold).