Move RDF Background before Wikidata Background
Wikidata already refers to RDF, so we should probably explain RDF first.
This will also let us explain wdt:P31 and wdt:P279 as opposed to
rdf:type and rdfs:subClassOf later, and move the definition of the
Wikidata-specific prefixes to the Wikidata section.

(The RDF Background also isn’t finished yet, but I’ll add that in
separate commits.)