Add example for RDF, replacing earlier list of prefixes
Instead of giving the prefixes in a terse list, including ones that
aren’t even related to RDF directly, introduce them gradually in the
text where they make sense. Also adds an example RDF graph in a
graphical representation and then in N-Triples syntax; it’s not the
greatest example, but it’s the best usable one I could think of to
include rdf:type and rdfs:subClassOf. (I did consider Beren and Lúthien
– one is a human, one is an elf, both are Children of Ilúvatar – but
ultimately that’s really too silly for this thesis.)