Add example of class cycle and explain why we deal with it
I *think* I got the history right now, but it’s confusing as hell. I
first thought that the edit to “mental representation” had introduced
the cycle, and that the cycle had been broken when that was reverted,
but that’s not the case, as far as I can tell.