Remove TODO
SPARQL is important because we use it (though we don’t need to get into
details, as this thesis never *shows* a single SPARQL query). Other RDF
technologies are not important because we don’t use them. (The ShEx
exporter does something JSON-LD-y, but I haven’t touched that part.)