Remove TODOs
The Toolforge section is fine as it is, I think. It *is* necessary –
otherwise the tool chapter is hard to understand – but there’s no need
to mention Cloud VPS, and I don’t think the chapter needs to have any
more content. It’s short, but that’s how it is.