Add CC BY 4.0 license
The csquotes package is added just to avoid an error from the doclicense
package – skimming through the documentation, it looks like it might
also have been useful on its own, but I think it’s too late to adopt to
using it now.

The hyperxmp package is used by the doclicense package to add licensing
information to the PDF document. For example, in Evince, you can press
Alt+Enter to see the document Properties, and this adds a Document
License tab to that dialog, with the usage terms and deed URL.