Remove TODO about type/datatype confusion in simplification
I haven’t looked into this in detail, but I tested a small sample graph
where the same resource was used as a class and as a datatype, and in
the resulting ShEx, both predicates used the resource as a class. So
if RDF2Graph is mixing up types and datatypes, that’s hardly our fault.
(Though, in fairness, in real-world graphs it’s probably extremely
unlikely that a datatype will also be used as a class.)