Fix layout of graph
Without the \vphantom{} added here (or some similar construct), the
ex:Dog node takes up slightly more vertical space than the ex:Human
node, due to the descender of the letter “g” (“ex:Human” has no letters
with descenders). This, in turn, causes the ex:Fala node to be slightly
below the ex:FDR node, which means that the arrows connecting them are
not quite perfectly horizontal. I’m not sure if this would’ve been
visible in print, but on a pixel-based display it was noticeable, and
it’s fairly easy to fix, so let’s do it.