Update Wikidata dump size
Give the number of triples, not lines (the number of lines approximates
the number of triples, but since the original count was on a Turtle
file, it’s not the same), and also update with the latest dump, and give
the size of the N-Triples dump instead of the Turtle dump, since it’s
more impressive (over one terabyte!).

Counted on Toolforge Grid with the following script:

    #!/bin/sh
    pv /public/dumps/public/wikidatawiki/entities/20181008/wikidata-20181008-all-BETA.nt.bz2 | bunzip2 | wc

(The pv is completely pointless, by the way, don’t bother with that if
you want to replicate this in the future – it doesn’t do anything if
stderr is not a pipe. There are options to fiddle with this, but I
couldn’t be bothered. Just be warned that this will take somewhere
between five and fifteen hours.)