Remove TODO about checking r² (R²) calculations
The formula looks correct enough to me, and I don’t have the time to
investigate it in more detail. R² can sometimes be negative, that’s just
how it is, I think. (This usually assumes that you’ve hamstrung your
regression function in some way – e. g. fixed an incorrect intercept for
a linear function – but in this case I think it might mean that gnuplot
is not that great at fitting nonlinear functions.)