Remove TODO about statement layout
It looks fine as a blockquote. It would probably also be possible to
make the \QL and \PL macros use the csquotes package to automatically
use single quotes when surrounde by double quotes, and then use an
inline quote instead of a blockquote, but I don’t have time to update
the whole thesis to csquotes and don’t want to do it just in this one
place, that feels inconsistent.