Fix caption location for Wikidata screenshots
All other floats in the thesis have the caption at the bottom, so it was
weird to have it at the top here. Much worse, though, this was
apparently counting TeX (LaTeX? the subcaption package? whatever) to
miscount the figures: though the subfigures were part of fig. 2.2,
references to them in the text used the counters 2.3a and 2.3b. It seems
strange that the placement of caption should affect this, but if it can
easily be fixed by moving the captions where they should be anyways for
different reasons, let’s do that.

This also repeats the figure caption below the second subfigure, since
it’s on a separate page.