Add Wikidata slides
These will be the backdrop for a very brief outline of Wikidata,
delivered orally. I’m using the Wikidata item for this thesis as the
example item, partly as a joke, but it also offers an opportunity to
quickly mention the constraints system and why it’s not sufficient:
constraints can tell us that the blog post I cite should have a title
statement, but they can’t tell us that the thesis item should cite other
items in the first place if no such statements were present.